+++ main.go
@@ -160,6 +160,22 @@ func main() {
 		}
 		env["GIT_REVISION"] = *pr.Head.SHA
 
+		// If PR is draft and triggered from the PR body don't run trigger and instead
+		// inform the user to add a comment
+		if *pr.Draft && env["COMMENT_ID"] == "" {
+			fmt.Println("PR is draft and was triggered from the opening comment, not triggering")
+
+			// TODO: Comment
+			_, _, err = ghClient.Issues.CreateComment(ctx, env["REPO_ORG"], env["REPO_NAME"], prNumber, &github.IssueComment{
+				Body: github.String("> [!NOTE]\n> As this is a draft PR no triggers from the PR body will be handled.\n> \n> If you'd like to trigger them while draft please add them as a PR comment."),
+			})
+			if err != nil {
+				fmt.Println("Failed to add PR comment", err)
+				os.Exit(1)
+			}
+			os.Exit(0)
+		}
+
 		// Get changed files
 		// TODO: Currently doesn't handle paging so limited to first 100 files but I'd like to refactor more before we introducing paging support
 		files, _, err := ghClient.PullRequests.ListFiles(ctx, env["REPO_ORG"], env["REPO_NAME"], prNumber, &github.ListOptions{PerPage: 100})